+++ src/main/java/com/example/service/ArtistService.java
@@ -17,7 +17,7 @@ public class ArtistService {
 
     public List<Artist> findAll() { return repository.findAll(); }
     public Artist save(Artist artist) { return repository.save(artist); }
-    public Optional<Artist> findById(Long id) { return repository.findById(id); }
+    public Optional<Artist> findById(Integer id) { return repository.findById(id); }
     public List<Artist> findByName(String name, org.springframework.data.domain.Sort sort) {
         throw new UnsupportedOperationException("Unimplemented method 'findByName'");
     }
